+++ SoftLayer/CLI/bandwidth/__init__.py
@@ -1,0 +1,2 @@
+"""Bandwidth."""
+# :license: MIT, see LICENSE for more details.
