+++ SoftLayer/CLI/bandwidth/pools_detail.py
@@ -1,4 +1,4 @@
-"""Get bandwidth pools."""
+"""Get bandwidth pool details."""
 # :license: MIT, see LICENSE for more details.
 import click
