+++ tests/CLI/modules/account_tests.py
@@ -165,17 +165,6 @@ class AccountCLITests(testing.TestCase):
         self.assert_called_with('SoftLayer_Account', 'getActiveVirtualLicenses')
         self.assert_called_with('SoftLayer_Account', 'getActiveAccountLicenses')
 
-    def test_bandwidth_pools(self):
-        result = self.run_command(['account', 'bandwidth-pools'])
-        self.assert_no_fail(result)
-        self.assert_called_with('SoftLayer_Account', 'getBandwidthAllotments')
-        self.assert_called_with('SoftLayer_Network_Bandwidth_Version1_Allotment', 'getObject')
-
-    def test_acccount_bandwidth_pool_detail(self):
-        result = self.run_command(['account', 'bandwidth-pools-detail', '123456'])
-        self.assert_no_fail(result)
-        self.assert_called_with('SoftLayer_Network_Bandwidth_Version1_Allotment', 'getObject')
-
     def test_acccount_provisioning_hook(self):
         result = self.run_command(['account', 'hooks'])
         self.assert_no_fail(result)
